Add imdb split info (#312)
* add split info

* add more docstring for imdb
