+++ paddlenlp/datasets/__init__.py
@@ -31,3 +31,4 @@ from .glue import *
 from .wmt14ende import *
 from .couplet import *
 from .yahoo_answer_100k import *
+from .imdb import *
